refactor: moving the wordify functions together
