+++ tools/generator-terraform/generator/resource/docs/helpers.go
@@ -1,6 +1,9 @@
 package docs
 
-import "strings"
+import (
+	"fmt"
+	"strings"
+)
 
 // wordifySegmentName takes an input PascalCased string and converts it to a more human-friendly variant
 // e.g. `applicationGroupName` -> `Application Group`
@@ -18,4 +21,36 @@ func wordifySegmentName(input string) string {
 	}
 
 	return strings.TrimPrefix(output, " ")
+}
+
+func wordifyTimeout(inMinutes int) string {
+	hours := inMinutes / 60
+	if hours > 0 {
+		var hoursText string
+		if hours > 1 {
+			hoursText = fmt.Sprintf("%d hours", hours)
+		} else {
+			hoursText = "1 hour"
+		}
+
+		minutesRemaining := inMinutes % 60
+		if minutesRemaining == 0 {
+			return hoursText
+		}
+
+		var minutesText string
+		if minutesRemaining > 1 {
+			minutesText = fmt.Sprintf("%d minutes", minutesRemaining)
+		} else {
+			minutesText = "1 minute"
+		}
+
+		return fmt.Sprintf("%s and %s", hoursText, minutesText)
+	}
+
+	if inMinutes > 1 {
+		return fmt.Sprintf("%d minutes", inMinutes)
+	}
+
+	return "1 minute"
 }
